requirements.txt
aiogram==3.17.0
python-dotenv==1.0.1
requests==2.32.3
matplotlib==3.10.0
seaborn==0.13.2
tka==0.6.4
httpx==1.0.0b0